Constraints

Constraints

GasModels.constraint_conserve_flow — Method.

This constraint is intended to ensure that flow is on direction through a node with degree 2 and no production or consumption

source

GasModels.constraint_conserve_flow_ne — Method.

This constraint is intended to ensure that flow is on direction through a node with degree 2 and no production or consumption

source

GasModels.constraint_flow_direction_choice — Method.

Constraint that states a flow direction must be chosen

source

GasModels.constraint_flow_direction_choice_ne — Method.

Constraint that states a flow direction must be chosen for new edges

source

GasModels.constraint_junction_flow_balance — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ls — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ne — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ne_ls — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_on_off_compressor_flow_direction — Method.

constraints on flow across compressors

source

GasModels.constraint_on_off_compressor_flow_direction_fixed_direction — Method.

constraints on flow across compressors when directions are constants

source

GasModels.constraint_on_off_compressor_flow_direction_ne — Method.

constraints on flow across compressors

source

GasModels.constraint_on_off_compressor_flow_direction_ne_fixed_direction — Method.

constraints on flow across compressors when the directions are constants

source

GasModels.constraint_on_off_compressor_ratios — Method.

enforces pressure changes bounds that obey compression ratios

source

GasModels.constraint_on_off_compressor_ratios_fixed_direction — Method.

on/off constraint for compressors when the flow direction is constant

source

GasModels.constraint_on_off_compressor_ratios_ne — Method.

constraints on pressure drop across control valves

source

GasModels.constraint_on_off_control_valve_flow_direction — Method.

constraints on flow across control valves

source

GasModels.constraint_on_off_control_valve_flow_direction_fixed_direction — Method.

constraints on flow across control valves when directions are constants

source

GasModels.constraint_on_off_control_valve_pressure_drop — Method.

constraints on pressure drop across control valves

source

GasModels.constraint_on_off_control_valve_pressure_drop_fixed_direction — Method.

constraints on pressure drop across control valves when directions are constants

source

GasModels.constraint_on_off_pipe_flow_direction — Method.

constraints on flow across pipes

source

GasModels.constraint_on_off_pipe_flow_direction_fixed_direction — Method.

constraints on flow across pipes where the directions are fixed

source

GasModels.constraint_on_off_pipe_flow_direction_ne — Method.

constraints on flow across pipes

source

GasModels.constraint_on_off_pipe_flow_direction_ne_fixed_direction — Method.

constraints on flow across pipes when directions are fixed

source

GasModels.constraint_on_off_pressure_drop — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_pressure_drop_fixed_direction — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_pressure_drop_ne — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_pressure_drop_ne_fixed_direction — Method.

constraints on pressure drop across pipes when the direction is fixed

source

GasModels.constraint_on_off_short_pipe_flow_direction — Method.

constraints on flow across short pipes

source

GasModels.constraint_on_off_short_pipe_flow_direction_fixed_direction — Method.

constraints on flow across short pipes when the directions are constants

source

GasModels.constraint_on_off_valve_flow_direction — Method.

constraints on flow across valves

source

GasModels.constraint_on_off_valve_flow_direction_fixed_direction — Method.

constraints on flow across valves when directions are constants

source

GasModels.constraint_on_off_valve_pressure_drop — Method.

constraints on pressure drop across valves

source

GasModels.constraint_parallel_flow — Method.

ensures that parallel lines have flow in the same direction

source

GasModels.constraint_parallel_flow_ne — Method.

ensures that parallel lines have flow in the same direction

source

GasModels.constraint_short_pipe_pressure_drop — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_sink_flow — Method.

Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

GasModels.constraint_sink_flow_ne — Method.

Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

GasModels.constraint_source_flow — Method.

Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source

GasModels.constraint_source_flow_ne — Method.

Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source