Network Formulations

Network Formulations

Type Hierarchy

We begin with the top of the hierarchy, where we can distinguish between gas flow models. Currently, there are two variations of the disjunctive form of the weymouth equations: The full non convex formulation and its conic relaxation.

AbstractMINLPForm <: AbstractGasFormulation
AbstractMISOCPForm <: AbstractGasFormulation

Gas Models

Each of these forms can be used as the type parameter for a GasModel:

MINLPGasModel = GenericGasModel(StandardMINLPForm)
MISOCPGasModel = GenericGasModel(StandardMISOCPForm)

For details on GenericGasModel, see the section on Gas Model.

User-Defined Abstractions

The user-defined abstractions begin from a root abstract like the AbstractGasFormulation abstract type, i.e.

AbstractMyFooForm <: AbstractGasFormulation

StandardMyFooForm <: AbstractFooForm
FooGasModel = GenericGasModel{StandardFooForm}

MINLP

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_fixed_direction — Method.

Weymouth equation with fixed direction variables

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with discrete direction variables for MINLP

source

GasModels.constraint_weymouth_ne_fixed_direction — Method.

Weymouth equation with fixed directions for MINLP

source

MISOCP

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_fixed_direction — Method.

Weymouth equation with fixed direction

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with discrete direction variables for MINLP

source

GasModels.constraint_weymouth_ne_fixed_direction — Method.

Weymouth equation with fixed direction

source

GasModels.variable_flux — Function.

source

GasModels.variable_flux_ne — Function.

source