Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in gas flow models. These methods should always be defined over "GenericGasModel".

GasModels.getstart — Function.

extracts the start value

source

GasModels.variable_compressor_ne — Function.

variables associated with building compressors

source

GasModels.variable_flux — Function.

variables associated with flux

source

GasModels.variable_flux_ne — Function.

variables associated with flux in expansion planning

source

GasModels.variable_load — Function.

variables associated with demand

source

GasModels.variable_pipe_ne — Function.

variables associated with building pipes

source

GasModels.variable_pressure_sqr — Function.

variables associated with pressure squared

source

GasModels.variable_production — Function.

variables associated with production

source

GasModels.variable_valve_operation — Function.

0-1 variables associated with operating valves

source