Constraints

Constraints

GasModels.constraint_compressor_flow — Method.

All constraints associated with flows through a compressor

source

GasModels.constraint_compressor_flow_ne — Method.

All constraints associated with flows through a compressor

source

GasModels.constraint_control_valve_flow — Method.

All constraints associated with flows through a control valve

source

GasModels.constraint_control_valve_flow_ne — Method.

All constraints associated with flows through a control valve

source

GasModels.constraint_exclusive_new_pipes — Method.

This function ensures at most one pipe in parallel is selected

source

GasModels.constraint_flow_direction_choice — Method.

Constraint that states a flow direction must be chosen

source

GasModels.constraint_flow_direction_choice_ne — Method.

Constraint that states a flow direction must be chosen for new edges

source

GasModels.constraint_junction_flow — Method.

All constraints associated with flows through at a junction

source

GasModels.constraint_junction_flow_balance — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ls — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ne — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_balance_ne_ls — Method.

standard flow balance equation where demand and production is fixed

source

GasModels.constraint_junction_flow_ls — Method.

All constraints associated with flows through at a junction with load shedding

source

GasModels.constraint_junction_flow_ne — Method.

All constraints associated with flows through at a junction with new pipe options

source

GasModels.constraint_junction_flow_ne_ls — Method.

All constraints associated with flows through at a junction with new pipe options

source

GasModels.constraint_new_compressor_flow_ne — Method.

All constraints associated with flows through a control valve

source

GasModels.constraint_new_compressor_ratios_ne — Method.

compressor rations have on off for direction and expansion

source

GasModels.constraint_new_pipe_flow_ne — Method.

All constraints associated with flows through a valve

source

GasModels.constraint_on_off_compressor_flow_ne — Method.

on/off constraints on flow across compressors for expansion variables

source

GasModels.constraint_on_off_compressor_ratios_ne — Method.

constraints on pressure drop across control valves

source

GasModels.constraint_on_off_pipe_flow_ne — Method.

on/off constraints on flow across pipes for expansion variables

source

GasModels.constraint_on_off_valve_pressure_drop — Method.

constraints on pressure drop across valves

source

GasModels.constraint_pipe_flow — Method.

All constraints associated with flows through a pipe

source

GasModels.constraint_pipe_flow_ne — Method.

All constraints associated with flows through a pipe

source

GasModels.constraint_short_pipe_flow — Method.

All constraints associated with flows through a short pipe

source

GasModels.constraint_short_pipe_flow_ne — Method.

All constraints associated with flows through a short pipe

source

GasModels.constraint_short_pipe_pressure_drop — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_valve_flow — Method.

All constraints associated with flows through a valve

source

GasModels.constraint_valve_flow_ne — Method.

All constraints associated with flows through a valve

source