Constraints

Constraints

Constraint Templates

Constraint templates help simplify data wrangling across multiple Gas Flow formulations by providing an abstraction layer between the network data and network constraint definitions. The constraint template's job is to extract the required parameters from a given network data structure and pass the data as named arguments to the Gas Flow formulations.

These templates should be defined over GenericGasModel and should not refer to model variables. For more details, see the files: core/constraint_template.jl and core/constraint.jl.

The convention is to define a function for each constraint. Utility functions are defined to collect together sets of constraints which are associated with individual components or capabilities of a component such as flow through a pipe.

Junction Constraints

Flow balance constraints

GasModels.constraint_junction_mass_flow_balance — Function.

Constraint: standard mass flow balance equation where demand and production are constants

source

Template: Constraints for mass flow balance equation where demand and production are constants

source

Flow balance constraints with load shedding

GasModels.constraint_junction_mass_flow_balance_ls — Function.

Constraint: standard flow balance equation where demand and production are variables

source

Template: Constraints for mass flow balance equation where demand and production is are a mix of constants and variables

source

Flow balance constraints for network expansion problems

GasModels.constraint_junction_mass_flow_balance_ne — Function.

Constraint: standard flow balance equation where demand and production are constants and there are expansion connections

source

Template: Constraints for mass flow balance equation where demand and production are constants and there are expansion connections

source

GasModels.constraint_junction_mass_flow_balance_ne_ls — Function.

Constraint: standard flow balance equation where demand and production are variables and there are expansion connections

source

Template: Constraints for mass flow balance equation where demand and production is are a mix of constants and variables and there are expansion connections

source

Direction On/off Constraints

GasModels.constraint_source_flow — Function.

Constraint: Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source

Template: Constraints for ensuring that at least one direction is set to take flow away from a junction (typically used on source nodes)

source

GasModels.constraint_sink_flow — Function.

Constraint: Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

Template: Constraints for ensuring that at least one direction is set to take flow away from a junction (typically used on sink nodes)

source

GasModels.constraint_conserve_flow — Function.

Constraint: This constraint is intended to ensure that flow is one direction through a node with degree 2 and no production or consumption

source

Template: Constraints to ensure that flow is the same direction through a node with degree 2 and no production or consumption

source

GasModels.constraint_source_flow_ne — Function.

Constraint: Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source

Template: Constraints for ensuring that at least one direction is set to take flow away from a junction (typically used on source nodes)

source

GasModels.constraint_sink_flow_ne — Function.

Constraint: Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

Template: Constraints for ensuring that at least one direction is set to take flow away from a junction (typically used on sink nodes)

source

GasModels.constraint_conserve_flow_ne — Function.

Constraint: This constraint is intended to ensure that flow is on direction through a node with degree 2 and no production or consumption for a node with expansion edges

source

Template: Constraints to ensure that flow is the same direction through a node with degree 2 and no production or consumption

source

Constraint collections

GasModels.constraint_junction_mass_flow — Function.

Template: All constraints associated with flows at a junction

source

Constraint: Constraints for computing mass flows at a junction

source

Constraint: Constraint for computing mass flow balance at node

source

GasModels.constraint_junction_mass_flow_directed — Function.

Template: All constraints associated with flows at a junction with some directed edges

source

Constraint: Constraints for computing mass flow balance at a junction when some edges are directed

source

GasModels.constraint_junction_mass_flow_ls — Function.

Template: All constraints associated with flows at a junction with variable production and consumption

source

Constraint: Constraints for computing mass flows at a junction with load variables

source

Constraint: Constraints for computing mass flow balance at node when injections are variables

source

GasModels.constraint_junction_mass_flow_ls_directed — Function.

Template: All constraints associated with flows through at a junction with variable production and consumption and some directed edges

source

Constraint: Constraints for computing mass flows at a junction when injections are variables and some edges are directed

source

GasModels.constraint_junction_mass_flow_ne — Function.

Template: All constraints associated with flows at a junction with new pipe options

source

Constraint: Constraints for computing mass flows at junctions with expansion variables

source

Constraint: Constraint for computing mass flow balance at node when there are expansion edges

source

GasModels.constraint_junction_mass_flow_ne_directed — Function.

Template: All constraints associated with flows through at a junction with expansion options and some directed edges

source

Constraint: Constraints for computing mass flows at a junction when there are expansion edges and some edges are directed

source

GasModels.constraint_junction_mass_flow_ne_ls — Function.

Template: All constraints associated with flows at a junction with variable production and consumption

source

Constraint: Constraints for computing mass flows at junctions with expansion and load variables

source

Constraint: Constraints for computing mass flow balance at node when there are expansion edges and variable injections

source

GasModels.constraint_junction_mass_flow_ne_ls_directed — Function.

Template: All constraints associated with flows through at a junction with expansion options, some directed edges, and variable production and consumption

source

Constraint: Constraints for computing mass flows at a junction when there are expansion edges, variable injections, and some edges are directed

source

Pipe Constraints

Weymouth's law constraints

GasModels.constraint_weymouth — Function.

Template: Weymouth equation for defining the relationship between pressure drop and flow across a pipe

source

Constraint: Weymouth equation–not applicable for MIP models

source

Weymouth equation with discrete direction variables

source

Weymouth equation for an undirected pipe

source

Direction on/off constraints

GasModels.constraint_on_off_pressure_drop — Function.

Constraint: Constraints which define pressure drop across a pipe when there are on/off direction variables

source

Template: Pressure drop across pipes with on/off direction variables

source

Missing docstring.

Missing docstring for constraint_on_off_pipe_flow. Check Documenter's build log for details.

GasModels.constraint_flow_direction_choice — Function.

Constraint: Constraints which state a flow direction must be chosen

source

Constraint: Choose one direction

source

GasModels.constraint_parallel_flow — Function.

Constraint: ensures that parallel lines have flow in the same direction

source

Template: Constraints which ensure that parallel lines have flow in the same direction

source

One way flow constraints

GasModels.constraint_pressure_drop_one_way — Function.

Constraint: constraints on pressure drop across where direction is constrained

source

Template: Constraint on pressure drop across a pipe where the flow is constrained to one direction as defined by data attributes yp and yn

source

Constraint: constraints on pressure drop across where direction is constrained

source

GasModels.constraint_pipe_flow_one_way — Function.

Constraint: constraint on flow across the pipe where direction is constrained

source

Template: Constraint on flow across a pipe where the flow is constrained to one direction as defined by data attributes yp and yn

source

Network Expansion Constraints

GasModels.constraint_weymouth_ne — Function.

Template: Weymouth equation for expansion pipes

source

Constraint: Weymouth equation–not applicable for MIP models–not applicable for MIP models

source

Weymouth equation for an undirected expansion pipe

source

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_on_off_pressure_drop_ne — Function.

Constraint: constraints on pressure drop across an expansion pipe with on/off direction variables

source

Template: Constraints on pressure drop across pipes with on/off direction variables

source

GasModels.constraint_on_off_pipe_flow_ne — Function.

Constraint: constraints on flow across an expansion pipe with on/off direction variables

source

Template: Constraints on flow across an expansion pipe with on/off direction variables

source

GasModels.constraint_flow_direction_choice_ne — Function.

Constraint: Constraint that states a flow direction must be chosen for expansion connections

source

Constraint: Choose one direction on expansion edges

source

GasModels.constraint_parallel_flow_ne — Function.

Constraint: ensures that parallel lines have flow in the same direction

source

Template: Constraints which ensure that parallel lines have flow in the same direction

source

GasModels.constraint_pressure_drop_ne_one_way — Function.

Constraint: Pressure drop across an expansion pipe when direction is constrained

source

Template: Constraint on pressure drop across an expansion pipe where the flow is constrained to one direction as defined by data attributes yp and yn

source

Constraint: Pressure drop across an expansion pipe when direction is constrained–not applicable for MIP models

source

GasModels.constraint_pipe_flow_ne_one_way — Function.

Constraint: Flow across an expansion pipe when direction is constrained

source

Template: Constraints on flow across an expansion pipe where the flow is constrained to one direction

source

Constraint collections

GasModels.constraint_pipe_flow — Function.

Template: All constraints associated with flows through a pipe

source

Constraint: Constraints which define flow across a pipe

source

Constraint: Constraints which define the pressure drop across a pipe when there are on/off direction variables

source

GasModels.constraint_pipe_flow_directed — Function.

Template: All constraints associated with flows through a pipe that has flow in one direction

source

Constraint: Constraints which define the flow across a pipe when the pipe is directed

source

GasModels.constraint_pipe_flow_ne — Function.

Template: All constraints associated with flows through an expansion pipe

source

Constraint: Constraints which define flow across a pipe

source

Constraint: All Constraints for an expansion pipe when there are on/off direction variables

source

GasModels.constraint_pipe_flow_ne_directed — Function.

Template: All constraints associated with flows through an expansion pipe which has flow in one direction

source

Constraint: Constraints for an expansion pipe where the direction of flow is constrained

source

Compressor Constraints

Direction On/off Constraints

GasModels.constraint_on_off_compressor_flow — Function.

Constraint: constraints on flow across a compressor with on/off direction variables

source

Template: constraints on flow across a compressor with on/off direction variables

source

GasModels.constraint_on_off_compressor_ratios — Function.

Constraint: enforces pressure changes bounds that obey compression ratios for a compressor with on/off direction variables

source

Template: constraints on compression ratios for compressors modeled with on/off direction variables

source

Network Expansion Constraints

GasModels.constraint_on_off_compressor_flow_ne — Function.

Constraint: constraints on flow across compressors with on/off direction variables

source

Template: constraints on flow across compressors where direction is controlled via on/off variables

source

GasModels.constraint_on_off_compressor_ratios_ne — Function.

Constraint: constraints on pressure drop across expansion compressors with on/off decision variables

source

Template: constraints on pressure drop on compressors where direction is controlled via on/off variables

source

Control Valve Constraints

Direction On/off Constraints

GasModels.constraint_on_off_control_valve_flow — Function.

constraints on flow across control valves that are undirected

source

Template: constraints on flow across control valves with on/off direction variables

source

Constraint: Constraints on flow across control valves with on/off direction variables

source

GasModels.constraint_on_off_control_valve_pressure_drop — Function.

Constraint: Constraints on pressure drop across control valves that have on/off direction variables

source

Template: constraints on pressure drop across control valves with on/off direction variables

source

Valve Constraints

Direction On/off Constraints

GasModels.constraint_on_off_valve_flow — Function.

constraints on flow across undirected valves

source

Template: constraints on flow across valves modeled with on/off direction variables

source

Constraint: Constraints on flow across valves with on/off direction variables

source

GasModels.constraint_valve_pressure_drop — Function.

Constraint: Constraints on pressure drop across valves where the valve can open or close

source

Template: Constraint on pressure drop across valves, where the valve may be closed or opened

source

Constraint: Constraints on pressure drop across valves where the valve can open or close–not applicable for MIP models

source

Short Pipes

Direction On/off Constraints

GasModels.constraint_on_off_short_pipe_flow — Function.

Constraint: Constraints on flow across a short pipe with on/off direction variables

source

Constraint: constraints on flow across a short pipe with on/off direction variables

source

GasModels.constraint_short_pipe_pressure_drop — Function.

Constraint: Constraint on pressure drop across a short pipe

source

Template: Constraint on pressure drop across a short pipe

source

Constraint: Constraint on pressure drop across a short pipe–not applicable for MIP models

source

Direction Cutting Constraints

Missing docstring.

Missing docstring for constraint_parallel_flow. Check Documenter's build log for details.

Network Expansion Constraints

Missing docstring.

Missing docstring for constraint_parallel_flow_ne. Check Documenter's build log for details.