Network Formulations

Network Formulations

Type Hierarchy

We begin with the top of the hierarchy, where we can distinguish between gas flow models. Currently, there are two variations of the weymouth equations, one where the directions of flux are known and one where they are unknown.

AbstractDirectedGasFormulation <: AbstractGasFormulation
AbstractUndirectedGasFormulation <: AbstractGasFormulation

Each of these have a disjunctive form of the weymouth equations: The full non convex formulation and its conic relaxation.

AbstractMINLPForm <: AbstractUndirectedGasFormulation
AbstractMISOCPForm <: AbstractUndirectedGasFormulation
AbstractMINLPDirectedForm <: AbstractDirectedGasFormulation
AbstractMISOCPDirectedForm <: AbstractDirectedGasFormulation

Gas Models

Each of these forms can be used as the type parameter for a GasModel, i.e.:

MINLPGasModel = GenericGasModel(StandardMINLPForm)
MISOCPGasModel = GenericGasModel(StandardMISOCPForm)

For details on GenericGasModel, see the section on Gas Model.

User-Defined Abstractions

The user-defined abstractions begin from a root abstract like the AbstractGasFormulation abstract type, i.e.

AbstractMyFooForm <: AbstractGasFormulation

StandardMyFooForm <: AbstractFooForm
FooGasModel = GenericGasModel{StandardFooForm}

Directed Models

GasModels.constraint_on_off_compressor_flow_direction — Method.

constraints on flow across compressors when directions are constants

source

GasModels.constraint_on_off_compressor_flow_direction_ne — Method.

constraints on flow across compressors when the directions are constants

source

GasModels.constraint_on_off_compressor_ratios — Method.

on/off constraint for compressors when the flow direction is constant

source

GasModels.constraint_on_off_control_valve_flow_direction — Method.

constraints on flow across control valves when directions are constants

source

GasModels.constraint_on_off_control_valve_pressure_drop — Method.

constraints on pressure drop across control valves when directions are constants

source

GasModels.constraint_on_off_pipe_flow_direction — Method.

constraints on flow across pipes where the directions are fixed

source

GasModels.constraint_on_off_pipe_flow_direction_ne — Method.

constraints on flow across pipes when directions are fixed

source

GasModels.constraint_on_off_pressure_drop — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_pressure_drop_ne — Method.

constraints on pressure drop across pipes when the direction is fixed

source

GasModels.constraint_on_off_short_pipe_flow_direction — Method.

constraints on flow across short pipes when the directions are constants

source

GasModels.constraint_on_off_valve_flow_direction — Method.

constraints on flow across valves when directions are constants

source

Undirected Models

GasModels.constraint_conserve_flow — Method.

This constraint is intended to ensure that flow is on direction through a node with degree 2 and no production or consumption

source

GasModels.constraint_conserve_flow_ne — Method.

This constraint is intended to ensure that flow is on direction through a node with degree 2 and no production or consumption

source

GasModels.constraint_on_off_compressor_flow_direction — Method.

constraints on flow across compressors

source

GasModels.constraint_on_off_compressor_flow_direction_ne — Method.

constraints on flow across compressors

source

GasModels.constraint_on_off_compressor_ratios — Method.

enforces pressure changes bounds that obey compression ratios

source

GasModels.constraint_on_off_control_valve_flow_direction — Method.

constraints on flow across control valves

source

GasModels.constraint_on_off_control_valve_pressure_drop — Method.

constraints on pressure drop across control valves

source

GasModels.constraint_on_off_pipe_flow_direction — Method.

constraints on flow across pipes

source

GasModels.constraint_on_off_pipe_flow_direction_ne — Method.

constraints on flow across pipes

source

GasModels.constraint_on_off_pressure_drop — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_pressure_drop_ne — Method.

constraints on pressure drop across pipes

source

GasModels.constraint_on_off_short_pipe_flow_direction — Method.

constraints on flow across short pipes

source

GasModels.constraint_on_off_valve_flow_direction — Method.

constraints on flow across valves

source

GasModels.constraint_parallel_flow — Method.

ensures that parallel lines have flow in the same direction

source

GasModels.constraint_parallel_flow_ne — Method.

ensures that parallel lines have flow in the same direction

source

GasModels.constraint_sink_flow — Method.

Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

GasModels.constraint_sink_flow_ne — Method.

Make sure there is at least one direction set to take flow to a junction (typically used on sink nodes)

source

GasModels.constraint_source_flow — Method.

Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source

GasModels.constraint_source_flow_ne — Method.

Make sure there is at least one direction set to take flow away from a junction (typically used on source nodes)

source

GasModels.variable_connection_direction — Method.

variables associated with direction of flow on the connections. yp = 1 imples flow goes from f_junction to t_junction. yn = 1 imples flow goes from t_junction to f_junction

source

GasModels.variable_connection_direction_ne — Method.

variables associated with direction of flow on the connections

source

MINLP

GasModels.constraint_weymouth — Method.

Weymouth equation with fixed direction variables

source

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with fixed directions for MINLP

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with discrete direction variables for MINLP

source

MISOCP

GasModels.constraint_weymouth — Method.

Weymouth equation with directed flow

source

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with fixed direction

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation with discrete direction variables for MINLP

source

GasModels.variable_flow — Method.

source

GasModels.variable_mass_flow — Method.

source

GasModels.variable_mass_flow_ne — Method.

source