Network Formulations

Network Formulations

Type Hierarchy

We begin with the top of the hierarchy, where we can distinguish between gas flow models. Currently, there are two variations of the weymouth equations, one where the directions of flux are known and one where they are unknown.

AbstractDirectedGasFormulation <: AbstractGasFormulation
AbstractUndirectedGasFormulation <: AbstractGasFormulation

Each of these have a disjunctive form of the weymouth equations: The full non convex formulation and its conic relaxation.

AbstractMINLPForm <: AbstractGasFormulation
AbstractMISOCPForm <: AbstractGasFormulation

Gas Models

Each of these forms can be used as the type parameter for a GasModel, i.e.:

MINLPGasModel = GenericGasModel(StandardMINLPForm)
MISOCPGasModel = GenericGasModel(StandardMISOCPForm)
MIPGasModel = GenericGasModel(StandardMIPForm)

For details on GenericGasModel, see the section on Gas Model.

User-Defined Abstractions

The user-defined abstractions begin from a root abstract like the AbstractGasFormulation abstract type, i.e.

AbstractMyFooForm <: AbstractGasFormulation

StandardMyFooForm <: AbstractFooForm
FooGasModel = GenericGasModel{StandardFooForm}

MINLP

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_weymouth_ne_one_way — Method.

Weymouth equation for directed expansion pipes

source

GasModels.constraint_weymouth_one_way — Method.

Weymouth equation with one way direction

source

MISOCP

GasModels.constraint_weymouth — Method.

Weymouth equation for an undirected pipe

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_weymouth_ne_one_way — Method.

Weymouth equation for expansion pipes with undirected expansion pipes

source

GasModels.constraint_weymouth_one_way — Method.

Weymouth equation with a pipe with one way flow

source

GasModels.variable_mass_flow — Method.

source

GasModels.variable_mass_flow_ne — Method.

source

MIP

GasModels.constraint_compressor_ratios — Method.

Constraint: Compressor ratio constraints on pressure differentials–not applicable for MIP models

source

GasModels.constraint_compressor_ratios_ne — Method.

Constraint: compressor ratios on a new compressor–not applicable for MIP models-not applicable for MIP models

source

GasModels.constraint_compressor_ratios_ne_one_way — Method.

Constraint: Pressure drop across an expansion compressor when direction is constrained-not applicable for MIP models

source

GasModels.constraint_compressor_ratios_one_way — Method.

Constraint: Compressor ratio when the flow direction is constrained–not applicable for MIP models

source

GasModels.constraint_control_valve_pressure_drop — Method.

constraints on pressure drop across control valves that are undirected–not applicable for MIP models

source

GasModels.constraint_control_valve_pressure_drop_one_way — Method.

Constraint: Pressure drop across a control valves when directions is constrained–not applicable for MIP models

source

GasModels.constraint_pressure_drop_ne_one_way — Method.

Constraint: Pressure drop across an expansion pipe when direction is constrained–not applicable for MIP models

source

GasModels.constraint_pressure_drop_one_way — Method.

Constraint: constraints on pressure drop across where direction is constrained

source

GasModels.constraint_short_pipe_pressure_drop — Method.

Constraint: Constraint on pressure drop across a short pipe–not applicable for MIP models

source

GasModels.constraint_valve_pressure_drop — Method.

Constraint: Constraints on pressure drop across valves where the valve can open or close–not applicable for MIP models

source

GasModels.constraint_weymouth — Method.

Constraint: Weymouth equation–not applicable for MIP models

source

GasModels.constraint_weymouth_ne — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for MIP models

source

GasModels.constraint_weymouth_ne_one_way — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for MIP models

source

GasModels.constraint_weymouth_one_way — Method.

Constraint: Weymouth equation with one way direction–not applicable for MIP models

source