Network Formulations

Network Formulations

Type Hierarchy

We begin with the top of the hierarchy, where we can distinguish between gas flow models. There are currently five formulations supported in GasModels. Two full non convex formulations and three relaxations.

AbstractNLPForm <: AbstractGasFormulation
AbstractMINLPForm <: AbstractGasFormulation
AbstractMISOCPForm <: AbstractGasFormulation
AbstractMIPForm <: AbstractGasFormulation
AbstractLPForm <: AbstractGasFormulation

Gas Models

Each of these forms can be used as the type parameter for a GasModel, i.e.:

NLPGasModel = GenericGasModel(StandardNLPForm)
MINLPGasModel = GenericGasModel(StandardMINLPForm)
MISOCPGasModel = GenericGasModel(StandardMISOCPForm)
MIPGasModel = GenericGasModel(StandardMIPForm)
LPGasModel = GenericGasModel(StandardLPForm)

For details on GenericGasModel, see the section on Gas Model.

User-Defined Abstractions

The user-defined abstractions begin from a root abstract like the AbstractGasFormulation abstract type, i.e.

AbstractMyFooForm <: AbstractGasFormulation

StandardMyFooForm <: AbstractFooForm
FooGasModel = GenericGasModel{StandardFooForm}

NLP

GasModels.constraint_compressor_ratios — Method.

enforces pressure changes bounds that obey compression ratios for an undirected compressor.

source

GasModels.constraint_compressor_ratios_ne — Method.

constraints on pressure drop across a compressor

source

GasModels.constraint_on_off_control_valve_pressure — Method.

constraints on pressure drop across control valves that are undirected

source

GasModels.constraint_weymouth — Method.

Weymouth equation with absolute value

source

GasModels.constraint_weymouth_directed — Method.

Weymouth equation with one way direction

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_weymouth_ne_directed — Method.

Weymouth equation for directed expansion pipes

source

MINLP

GasModels.constraint_weymouth — Method.

Weymouth equation with discrete direction variables

source

GasModels.constraint_weymouth_directed — Method.

Weymouth equation with one way direction

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_weymouth_ne_directed — Method.

Weymouth equation for directed expansion pipes

source

MISOCP

GasModels.constraint_weymouth — Method.

Weymouth equation for an undirected pipe

source

GasModels.constraint_weymouth_directed — Method.

Weymouth equation with a pipe with one way flow

source

GasModels.constraint_weymouth_ne — Method.

Weymouth equation for an undirected expansion pipe

source

GasModels.constraint_weymouth_ne_directed — Method.

Weymouth equation for expansion pipes with undirected expansion pipes

source

GasModels.variable_mass_flow — Method.

source

GasModels.variable_mass_flow_ne — Method.

source

MIP

GasModels.constraint_compressor_ratios — Method.

Constraint: Compressor ratio constraints on pressure differentials–not applicable for MIP models

source

GasModels.constraint_compressor_ratios_directed — Method.

Constraint: Compressor ratio when the flow direction is constrained–not applicable for MIP models

source

GasModels.constraint_compressor_ratios_ne — Method.

Constraint: compressor ratios on a new compressor–not applicable for MIP models-not applicable for MIP models

source

GasModels.constraint_compressor_ratios_ne_directed — Method.

Constraint: Pressure drop across an expansion compressor when direction is constrained-not applicable for MIP models

source

GasModels.constraint_on_off_control_valve_pressure — Method.

constraints on pressure drop across control valves that are undirected–not applicable for MIP models

source

GasModels.constraint_on_off_control_valve_pressure_directed — Method.

Constraint: Pressure drop across a control valves when directions is constrained–not applicable for MIP models

source

GasModels.constraint_on_off_valve_pressure — Method.

Constraint: Constraints on pressure drop across valves where the valve can open or close–not applicable for MIP models

source

GasModels.constraint_pipe_pressure — Method.

Constraint: Constraints which define pressure drop across a pipe

source

GasModels.constraint_pipe_pressure_directed — Method.

Constraint: constraints on pressure drop across where direction is constrained

source

GasModels.constraint_pipe_pressure_ne — Method.

Constraint: constraints on pressure drop across an expansion pipe

source

GasModels.constraint_pressure_ne_directed — Method.

Constraint: Pressure drop across an expansion pipe when direction is constrained–not applicable for MIP models

source

GasModels.constraint_short_pipe_pressure — Method.

Constraint: Constraint on pressure drop across a short pipe–not applicable for MIP models

source

GasModels.constraint_weymouth — Method.

Constraint: Weymouth equation–not applicable for MIP models

source

GasModels.constraint_weymouth_directed — Method.

Constraint: Weymouth equation with one way direction–not applicable for MIP models

source

GasModels.constraint_weymouth_ne — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for MIP models

source

GasModels.constraint_weymouth_ne_directed — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for MIP models

source

LP

GasModels.constraint_compressor_ratios — Method.

Constraint: Compressor ratio constraints on pressure differentials–not applicable for LP models

source

GasModels.constraint_compressor_ratios_directed — Method.

Constraint: Compressor ratio when the flow direction is constrained–not applicable for LP models

source

GasModels.constraint_compressor_ratios_ne — Method.

Constraint: compressor ratios on a new compressor–not applicable for MIP models-not applicable for LP models

source

GasModels.constraint_compressor_ratios_ne_directed — Method.

Constraint: Pressure drop across an expansion compressor when direction is constrained-not applicable for LP models

source

GasModels.constraint_on_off_control_valve_pressure — Method.

constraints on pressure drop across control valves that are undirected–not applicable for LP models

source

GasModels.constraint_on_off_control_valve_pressure_directed — Method.

Constraint: Pressure drop across a control valves when directions is constrained–not applicable for LP models

source

GasModels.constraint_on_off_valve_pressure — Method.

Constraint: Constraints on pressure drop across valves where the valve can open or close–not applicable for LP models

source

GasModels.constraint_pipe_pressure — Method.

Constraint: Constraints which define pressure drop across a pipe

source

GasModels.constraint_pipe_pressure_directed — Method.

Constraint: constraints on pressure drop across where direction is constrained

source

GasModels.constraint_pipe_pressure_ne — Method.

Constraint: constraints on pressure drop across an expansion pipe

source

GasModels.constraint_pressure_drop_ne_directed — Method.

Constraint: Pressure drop across an expansion pipe when direction is constrained–not applicable for LP models

source

GasModels.constraint_short_pipe_pressure — Method.

Constraint: Constraint on pressure drop across a short pipe–not applicable for LP models

source

GasModels.constraint_weymouth — Method.

Constraint: Weymouth equation–not applicable for LP models

source

GasModels.constraint_weymouth_directed — Method.

Constraint: Weymouth equation with one way direction–not applicable for LP models

source

GasModels.constraint_weymouth_ne — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for LP models

source

GasModels.constraint_weymouth_ne_directed — Method.

Constraint: Weymouth equation–not applicable for MIP models–not applicable for LP models

source

GasModels.variable_valve_operation — Method.

continous relaxation of variables associated with operating valves

source