Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in gas flow models. These methods should always be defined over "GenericGasModel".

GasModels.getstart — Function.

extracts the start value

source

GasModels.variable_compressor_ne — Method.

variables associated with building compressors

source

GasModels.variable_load_mass_flow — Method.

variables associated with demand

source

GasModels.variable_mass_flow — Method.

variables associated with mass flow

source

GasModels.variable_mass_flow_ne — Method.

variables associated with mass flow in expansion planning

source

GasModels.variable_pipe_ne — Method.

variables associated with building pipes

source

GasModels.variable_pressure_sqr — Method.

variables associated with pressure squared

source

GasModels.variable_production_mass_flow — Method.

variables associated with production

source

GasModels.variable_valve_operation — Method.

0-1 variables associated with operating valves

source