Variables

We provide the following methods to provide a compositional approach for defining common variables used in gas flow models. These methods should always be defined over "AbstractGasModel".

GasModels.get_compressor_pressure_aux — Method

Support function for getting a one off auxilleary pressure variable

GasModels.get_compressor_y — Method

Support function for getting a one off y direction variable

GasModels.get_ne_compressor_pressure_aux — Method

Support function for getting a one off auxilleary pressure variable

GasModels.get_ne_compressor_y — Method

Support function for getting a one off y direction variable

GasModels.variable_compressor_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_compressor_direction_ne — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_compressor_mass_flow — Function

variables associated with mass flow in compressors

GasModels.variable_compressor_mass_flow_ne — Function

variables associated with mass flow in compressors in expansion planning

GasModels.variable_compressor_minpower_proxy — Function

Variable Set: variables associated with proxy for minimizing compression power

GasModels.variable_compressor_ne — Function

variables associated with building compressors

GasModels.variable_compressor_ratio_sqr — Function

Variable Set: variables associated with compression ratios

GasModels.variable_compressor_ratio_sqr_ne — Function

Variable Set: variables associated with compression ratios

GasModels.variable_connection_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_connection_direction_ne — Function

variables associated with direction of flow on new connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction

GasModels.variable_flow — Function

Variable Set: Define variables needed for modeling flow across connections

GasModels.variable_flow_ne — Function

Variable Set: Define variables needed for modeling flow across connections that are expansions

GasModels.variable_load_mass_flow — Function

variables associated with demand

GasModels.variable_loss_resistor_direction — Function

variables associated with direction of flow on on lossresistors. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to f_junction

GasModels.variable_loss_resistor_mass_flow — Function

variables associated with mass flow in loss_resistors

GasModels.variable_mass_flow — Function

all variables associated with mass flow

GasModels.variable_mass_flow_ne — Function

all variables associated with mass flow in expansion planning

GasModels.variable_on_off_operation — Function

0-1 variables associated with operating edge components

GasModels.variable_pipe_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_pipe_direction_ne — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_pipe_mass_flow — Function

variables associated with mass flow in pipes

GasModels.variable_pipe_mass_flow_ne — Function

variables associated with mass flow in pipes in expansion planning

GasModels.variable_pipe_ne — Function

variables associated with building pipes

GasModels.variable_pressure — Function

variables associated with (nonsquared) pressure

GasModels.variable_pressure_sqr — Function

variables associated with pressure squared

GasModels.variable_production_mass_flow — Function

variables associated with production

GasModels.variable_regulator_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_regulator_mass_flow — Function

variables associated with mass flow in regulators

GasModels.variable_regulator_on_off_operation — Function

0-1 variables associated with operating regulators

GasModels.variable_resistor_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_resistor_mass_flow — Function

variables associated with mass flow in resistors

GasModels.variable_short_pipe_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_short_pipe_mass_flow — Function

variables associated with mass flow in short pipes

GasModels.variable_storage — Function

Variable Set: Define variables needed for modeling flow across storage

GasModels.variable_storage_mass_flow — Function

variables associated with storage flows

GasModels.variable_transfer_mass_flow — Function

variables associated with transfer

GasModels.variable_valve_direction — Function

variables associated with direction of flow on the connections. y = 1 imples flow goes from fjunction to tjunction. y = 0 imples flow goes from tjunction to fjunction. O flow can have y = 0 or 1

GasModels.variable_valve_mass_flow — Function

variables associated with mass flow in valves

GasModels.variable_valve_on_off_operation — Function

0-1 variables associated with operating valves