Variables

Variables

We provide the following methods to provide a compositional approach for defining common variables used in gas flow models. These methods should always be defined over "GenericGasModel".

GasModels.getstart — Function.

extracts the start value

source

GasModels.variable_compressor_ne — Function.

variables associated with building compressors

source

GasModels.variable_connection_direction — Method.

variables associated with direction of flow on the connections. yp = 1 imples flow goes from fjunction to tjunction. yn = 1 imples flow goes from tjunction to fjunction

source

GasModels.variable_connection_direction_ne — Method.

variables associated with direction of flow on the connections yp = 1 imples flow goes from fjunction to tjunction. yn = 1 imples flow goes from tjunction to fjunction

source

GasModels.variable_flow — Function.

Variable Set: Define variables needed for modeling flow across connections

source

GasModels.variable_flow_directed — Function.

Variable Set: Define variables needed for modeling flow across connections where some flows are directionally constrained

source

GasModels.variable_flow_ne — Function.

Variable Set: Define variables needed for modeling flow across connections that are expansions

source

GasModels.variable_flow_ne_directed — Function.

Variable Set: Define variables needed for modeling flow across connections that are expansions and some flows are directionally constrained

source

GasModels.variable_load_mass_flow — Function.

variables associated with demand

source

GasModels.variable_mass_flow — Function.

variables associated with mass flow

source

GasModels.variable_mass_flow_ne — Function.

variables associated with mass flow in expansion planning

source

GasModels.variable_pipe_ne — Function.

variables associated with building pipes

source

GasModels.variable_pressure_sqr — Method.

variables associated with pressure squared

source

GasModels.variable_production_mass_flow — Function.

variables associated with production

source

GasModels.variable_valve_operation — Method.

0-1 variables associated with operating valves

source